move mixtral moe layer back
